+++ usr.sbin/bhyve/Makefile
@@ -54,6 +54,7 @@ SRCS=	\
 	tpm_emul_passthru.c	\
 	tpm_intf_crb.c		\
 	tpm_ppi_qemu.c		\
+	uart_backend.c		\
 	uart_emul.c		\
 	usb_emul.c		\
 	usb_mouse.c		\
